requirements.txt
groq>=0.11.0
python-dotenv>=1.0.0
rich>=13.7.0
gradio>=4.0.0
